requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
cachetools==5.3.2
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
daphne==4.1.0
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-base64field==1.0
django-braces==1.15.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-ckeditor==6.7.1
django-cleanup==8.1.0
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-filter==23.5
django-js-asset==2.2.0
django-oauth-toolkit==0.12.0
django-phonenumber-field==7.3.0
django-resized==1.0.2
django-rest-framework-social-oauth2==1.0.0
django-rest-registration==0.8.3
django-reversion==5.0.12
django-timezone-field==6.1.0
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
drf-extra-fields==3.7.0
drf-social-oauth2==2.1.3
drf-writable-nested==0.7.0
drf-yasg==1.21.7
ecdsa==0.18.0
facebook-sdk==3.1.0
filetype==1.2.0
google-auth==2.28.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.4
kombu==5.3.7
Markdown==3.5.2
msgpack==1.0.8
oauth2_provider==0.0
oauthlib==2.0.1
packaging==23.2
phonenumbers==8.13.30
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.0.0
rpds-py==0.18.0
rsa==4.9
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
swagger-spec-validator==3.0.3
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
